+++ src/shared/modules/user/index.ts
@@ -1,4 +1,4 @@
 export { DefaultUserService } from './default-user-service.js';
-export { CreateUserDto } from './dto/create-user.dto.js';
+export { CreateUserDto } from './dto/create-user-dto.js';
 export { createUserContainer } from './user-container.js';
 export { UserEntity, UserModel } from './user-entity.js';
