+++ src/shared/modules/user/user-service.interface.ts
@@ -1,6 +1,6 @@
 import { DocumentType } from '@typegoose/typegoose';
 
-import { CreateUserDto } from './dto/create-user.dto.js';
+import { CreateUserDto } from './dto/create-user-dto.js';
 import { UserEntity } from './user-entity.js';
 
 export type UserEntityDocument = DocumentType<UserEntity>;
